* fixed generics bound paramter issue
git-svn-id: svn+ssh://svn.synyx.de/var/svn/synyx/opensource/hades/trunk@3092 5a64d73e-33d6-4ccc-9058-23f8668ecac9

+++ src/test/java/org/synyx/hades/dao/orm/support/GenericDaoFactoryBeanUnitTest.java
@@ -206,7 +206,7 @@ public class GenericDaoFactoryBeanUnitTest {
      * 
      * @author Oliver Gierke - gierke@synyx.de
      */
-    private interface SampleCustomDao extends GenericDao<User, Long> {
+    private interface SampleCustomDao extends GenericDao<User, Integer> {
 
         void someSampleMethod();
     }
